+++ CHANGELOG.rst
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+**Version 0.1.2** *(March 26, 2010)*
+
+* Normalized slide paths.
+* Configurable file encoding for slides.
+* A static "offline" version of a presentation can now be generated through the
+  command-line tool.
+
 **Version 0.1.1** *(March 16, 2010)*
 
 * Fixed packaging problem.